requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
numpy
tensorflow
tqdm
gymnasium
python-dotenv
redis